Separate utils function from model base functions
